research/requirements-research.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
huggingface_hub>=0.19.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Experiment Tracking
wandb>=0.15.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Interpretability
shap>=0.42.0
lime>=0.2.0.1
captum>=0.6.0  # For Integrated Gradients (text highlighting)

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
python-dotenv>=1.0.0

# Web Application
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0
